% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo-ape-info.r, R/data-management.r
\name{Nann.aphylo_estimates}
\alias{Nann.aphylo_estimates}
\alias{aphylo-methods}
\alias{plot.aphylo}
\alias{print.aphylo}
\alias{summary.aphylo}
\title{Plot and print methods for \code{aphylo} objects}
\usage{
\method{Nann}{aphylo_estimates}(phy)

\method{plot}{aphylo}(x, y = NULL, prop = 0.15, rect.args = list(),
  ...)

\method{print}{aphylo}(x, ...)

\method{summary}{aphylo}(object, ...)
}
\arguments{
\item{x}{An object of class \code{aphylo}.}

\item{y}{Ignored.}

\item{prop}{Numeric scalar between 0 and 1. Proportion of the device that the
annotations use in \code{plot.aphylo}.}

\item{rect.args}{List of arguments passed to \link[graphics:rect]{graphics::rect}.}

\item{...}{Further arguments passed to \link[ape:plot.phylo]{ape::plot.phylo}.}

\item{object}{An object of class \code{aphylo}.}
}
\value{
In the case of \code{plot.aphylo}, \code{NULL}.

In the case of \code{print.aphylo}, the same \code{aphylo} object
invisible.
}
\description{
Plot and print methods for \code{aphylo} objects
}
\details{
The \code{plot.aphylo} function is a wrapper of \link[ape:plot.phylo]{ape::plot.phylo}.
}
\seealso{
Other aphylo methods: \code{\link{aphylo-class}}
}
\concept{aphylo methods}
