% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo-package.R
\docType{package}
\name{aphylo-package}
\alias{aphylo-package}
\title{Statistical Inference in Annotated Phylogenetic Trees}
\description{
Statistical Inference in Annotated Phylogenetic Trees
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USCbiostats/aphylo}
  \item Report bugs at \url{https://github.com/USCbiostats/aphylo/issues}
}

}
\author{
\strong{Maintainer}: George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})

Other contributors:
\itemize{
  \item National Cancer Institute (NCI) (Grant Number 5P01CA196569-02) [funder]
  \item USC Biostatistics [copyright holder]
}

}
\section{How to cite}{
  If you use \pkg{aphylo} in published work, please cite it. Run
  \code{citation("aphylo")} in R for the full entry.
}
