% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r, R/package.r
\docType{package}
\name{leafs.phylo}
\alias{leafs.phylo}
\alias{leafs.po_tree}
\alias{leafs.aphylo}
\alias{aphylo}
\alias{aphylo-package}
\title{Statistical Inference in Annotated Phylogenetic Trees}
\usage{
\method{leafs}{phylo}(x, ...)

\method{leafs}{po_tree}(x, ...)

\method{leafs}{aphylo}(x, ...)
}
\description{
Statistical Inference in Annotated Phylogenetic Trees
}
