% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.r
\name{aphylo_cv}
\alias{aphylo_cv}
\alias{aphylo_cv.formula}
\title{Leave-one-out Cross Validation}
\usage{
aphylo_cv(...)

\method{aphylo_cv}{formula}(model, ...)
}
\arguments{
\item{...}{Further arguments passed to the method.}

\item{model}{As passed to \link{aphylo_mcmc}.}
}
\value{
An object of class \code{aphylo_cv} with the following components:
\itemize{
\item \code{pred_out} Out of sample prediction.
\item \code{expected} Expected annotations
\item \code{call} The call
\item \code{ids} Integer vector with the ids of the leafs used in the loo process.
}
}
\description{
Leave-one-out Cross Validation
}
