% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo_hier.R
\name{aphylo_hier}
\alias{aphylo_hier}
\title{Fit a Hierarchical aphylo model}
\usage{
aphylo_hier(
  model,
  params,
  classes,
  ...,
  params0 = NULL,
  hyper_params = NULL,
  env = parent.frame(),
  verbose = TRUE,
  multicore = FALSE,
  nchains = 1L,
  use_optim = TRUE,
  priors = NULL,
  check_informative = NULL,
  reduced_pseq = NULL
)
}
\arguments{
\item{model}{A model as specified in \link{aphylo-model}.}

\item{params}{A vector of length 7 with initial parameters. In particular
\code{psi[1]}, \code{psi[2]}, \code{mu[1]}, \code{mu[2]}, \code{eta[1]}, \code{eta[2]} and \code{Pi}.}

\item{classes}{An integer vector of length equal to the number of trees in
the model.}

\item{..., multicore, nchains}{passed to \link[fmcmc:MCMC]{fmcmc::MCMC}}

\item{params0}{Starting parameters.}

\item{hyper_params}{If not specified, the function sets as initial hyper
parameters equal to 10.}

\item{env}{Environment where to evaluate \code{model}.}

\item{verbose}{Logical scalar. When \code{TRUE} prints information.}

\item{use_optim}{Logical, When true uses \link[stats:optim]{stats::optim} as a starting point.}

\item{priors}{A function to be used as prior for the model (see \link{bprior}).}

\item{check_informative}{Logical scalar. When \code{TRUE} the algorithm
stops with an error when the annotations are uninformative (either 0s or 1s).}

\item{reduced_pseq}{Logical. When \code{TRUE} it will use a reduced peeling sequence
in which it drops unannotated leafs. If the model includes \code{eta} this is set
to \code{FALSE}.}
}
\value{
An object of class \link{aphylo_estimates}.
}
\description{
Fit a Hierarchical aphylo model
}
\details{
The parameters \code{priors}, \code{check_informative}, and \code{reduced_pseq}
are silently ignored in this function.
}
\seealso{
Other parameter estimation: 
\code{\link{aphylo_mcmc}()},
\code{\link{aphylo_mle}()}
}
\concept{parameter estimation}
