% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo_mcmc.R
\name{aphylo_mcmc}
\alias{aphylo_mcmc}
\alias{MCMC}
\title{Model estimation using Markov Chain Monte Carlo}
\usage{
aphylo_mcmc(
  model,
  params,
  priors = uprior(),
  control = list(),
  check_informative = getOption("aphylo_informative", FALSE),
  reduced_pseq = getOption("aphylo_reduce_pseq", TRUE)
)
}
\arguments{
\item{model}{A model as specified in \link{aphylo-model}.}

\item{params}{A vector of length 7 with initial parameters. In particular
\code{psi[1]}, \code{psi[2]}, \code{mu[1]}, \code{mu[2]}, \code{eta[1]}, \code{eta[2]} and \code{Pi}.}

\item{priors}{A function to be used as prior for the model (see \link{bprior}).}

\item{control}{A list with parameters for the optimization method (see
details).}

\item{check_informative}{Logical scalar. When \code{TRUE} the algorithm
stops with an error when the annotations are uninformative (either 0s or 1s).}

\item{reduced_pseq}{Logical. When \code{TRUE} it will use a reduced peeling sequence
in which it drops unannotated leafs. If the model includes \code{eta} this is set
to \code{FALSE}.}
}
\value{
An object of class \link{aphylo_estimates}.
}
\description{
The function is a wrapper of \code{\link[fmcmc:MCMC]{fmcmc::MCMC()}}.
}
\details{
Methods \code{\link[base:print]{base::print()}}, \code{\link[base:summary]{base::summary()}}, \link[stats:coef]{stats::coef}, \code{\link[stats:window]{stats::window()}},
\code{\link[stats:vcov]{stats::vcov()}}, \code{\link[stats:logLik]{stats::logLik()}}, \link[=predict.aphylo_estimates]{predict()},
and the various ways to query features of the trees via \link[ape:Ntip]{Ntip()}
are available post estimation.
}
\examples{
# Using the MCMC ------------------------------------------------------------

\dontrun{

set.seed(1233)
# Simulating a tree
tree <- sim_tree(200)

# Simulating functions
dat <- raphylo(
  tree = tree,
  psi  = c(.01, .03),
  mu_d = c(.05, .02),
  Pi   = .5
)

# Running the MCMC
set.seed(1231)

ans_mcmc <- aphylo_mcmc(
  dat ~ mu_d + psi + eta + Pi,
  control = list(nsteps = 2e5, burnin=1000, thin=200)
)
}
}
\seealso{
Other parameter estimation: 
\code{\link{aphylo_mle}()}
}
\concept{parameter estimation}
