% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{as.apephylo}
\alias{as.apephylo}
\alias{as.aphylo}
\alias{as.apephylo.aphylo}
\alias{as.apephylo.po_tree}
\alias{as_ape_tree}
\title{Coercing into \code{phylo} objects of the \pkg{ape} package.}
\usage{
as.apephylo(x, ...)

\method{as.apephylo}{aphylo}(x, ...)

\method{as.apephylo}{po_tree}(x, ...)

as_ape_tree(edges)
}
\arguments{
\item{x}{An object of class \code{\link[=as_po_tree]{po_tree}} or
\code{\link[=new_aphylo]{aphylo}}.}

\item{...}{Ignored.}

\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}
}
\value{
An object of class \code{\link[ape:as.phylo]{phylo}}

An integer matrix of the same dimmension as \code{edges} with the following
two aditional attributes:
\item{labels}{Named integer vector of size \code{n}. Original labels of the edgelist
where the first \code{n} are leaf nodes, \code{n+1} is the root node, and the reminder
are the internal nodes.}
\item{isleaf}{Logical vector of size \code{n}. \code{TRUE} when the node is a leaf node.}
}
\description{
\code{as.apephylo} coerces objects to \code{\link[ape:as.phylo]{phylo}} objects
from the \pkg{ape} package. These have several methods including visualization
methods that can be useful.

\code{as_ape_tree} is the powerhorse of \code{\link{as.apephylo}}.

\code{}
}
\examples{

# A simple example ----------------------------------------------------------
set.seed(1231)

# Generating a random tree: Observe that the nodes are coded such that
# for all (i,j) that are offspring and parent i > j. This is a requirement
# of the peeling algorithm
edges <- sim_tree(30)

# In the ape package, nodes are labeled such that leafs are from 1 to n
# root node is n+1 and interior nodes can have whatever label.
ape_edges <- as_ape_tree(edges)

# And further, is easy to go back to the original code
edges_org <- ape_edges
edges_org[] <- attr(ape_edges, "labels")[ape_edges[]]

# Comparing the three versions
head(edges)
head(ape_edges)
head(edges_org)

}
\seealso{
Other Data management functions: \code{\link{aphylo-class}},
  \code{\link{as_po_tree}}
}
