% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.r
\name{as.phylo.aphylo}
\alias{as.phylo.aphylo}
\title{Coercing into \code{phylo} objects of the \pkg{ape} package.}
\usage{
\method{as.phylo}{aphylo}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=as_po_tree:po_tree]{=as_po_tree::po_tree()}} or
\code{\link[=new_aphylo:aphylo]{=new_aphylo::aphylo()}}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{\link[ape:as.phylo:phylo]{ape:as.phylo::phylo()}}
}
\description{
\code{as.apephylo} coerces objects to \code{\link[ape:as.phylo:phylo]{ape:as.phylo::phylo()}} objects
from the \pkg{ape} package. These have several methods including visualization
methods that can be useful.
}
\seealso{
Other Data management functions: \code{\link{aphylo-class}},
  \code{\link{prune}}
}
\concept{Data management functions}
