% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.r
\name{as.phylo.matrix}
\alias{as.phylo.matrix}
\title{Extensions to the \code{as.phylo} function}
\usage{
\method{as.phylo}{matrix}(x, edge.length = NULL, root.edge = NULL, ...)
}
\arguments{
\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}
}
\value{
An integer matrix of the same dimmension as \code{edges} with the following
aditional attribute:
\item{labels}{Named integer vector of size \code{n}. Original labels of the edgelist
where the first \code{n} are leaf nodes, \code{n+1} is the root node, and the reminder
are the internal nodes.}
}
\description{
This function takes an edgelist and recodes (relabels) the nodes following
\CRANpkg{ape}'s coding convention. \code{as_ape_tree} is the powerhorse of
\code{\link[=as.apephylo]{as.apephylo()}}.
}
\examples{

# A simple example ----------------------------------------------------------
set.seed(1231)

# Generating a random tree: Observe that the nodes are coded such that
# for all (i,j) that are offspring and parent i > j. This is a requirement
# of the peeling algorithm
edges <- sim_tree(30)

# In the ape package, nodes are labeled such that leafs are from 1 to n
# root node is n+1 and interior nodes can have whatever label.
ape_edges <- as_ape_tree(edges)

# And further, is easy to go back to the original code
edges_org <- ape_edges
edges_org[] <- attr(ape_edges, "labels")[ape_edges[]]

# Comparing the three versions
head(edges)
head(ape_edges)
head(edges_org)

}
