% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{as.phylo}
\alias{as.phylo}
\alias{as.phylo.default}
\title{Coercing into \code{phylo} objects of the \pkg{ape} package.}
\usage{
as.phylo(x, ...)

\method{as.phylo}{default}(x, ...)
}
\arguments{
\item{x}{An object of class \code{phylo_tree} or \code{aphylo}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{\link[ape:as.phylo]{phylo}}
}
\description{
\code{phylo} objects have several methods such as \code{plot}, \code{print}
and \code{summary}.
}
\seealso{
Other Data management functions: \code{\link{new_aphylo}},
  \code{\link{recode_vertices}}
}
