% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{bprior}
\alias{bprior}
\alias{uprior}
\title{Default priors for \link{aphylo_mcmc}}
\usage{
bprior(shape1 = 1, shape2 = 9, ...)

uprior()
}
\arguments{
\item{shape1, shape2, ...}{Arguments passed to \link[stats:Beta]{stats::dbeta}}
}
\value{
In the case of \code{bprior}, a wrapper of the function \link[stats:Beta]{stats::dbeta}.
\code{uprior} returns a function \code{function(p) 1} (the uniform prior)
}
\description{
Convenient wrappers to be used with the \code{aphylo} estimation methods.
}
\examples{
bprior(1, 9)
uprior()
}
