% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{experiment}
\alias{experiment}
\title{Experimental Data}
\format{A data frame with 365 rows and 4 variables:
\describe{
  \item{f01}{State of function 1.}
  \item{f02}{State of function 1.}
  \item{f03}{State of function 1.}
  \item{LeafId}{Integer, ID of the leaf.}
}}
\source{
BiostatsUSC
}
\description{
A dataset containing 3 functional state of the leaf nodes. Each
function can have either 0 (unactive), 1 (active) or 9 (n/a).
}
