% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo-package.R
\name{fakeexperiment}
\alias{fakeexperiment}
\title{Fake Experimental Data}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{f1}{State of function 1.}
\item{f2}{State of function 1.}
\item{LeafId}{Integer, ID of the leaf.}
}
}
\source{
BiostatsUSC
}
\description{
A fake dataset containing 2 functional state of the leaf nodes. Each
function can have either 0 (unactive), 1 (active) or 9 (n/a).
This dataset is inteded for testing only.
}
