% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo-package.R
\name{faketree}
\alias{faketree}
\title{Fake Phylogenetic Tree}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{NodeId}{Integer, ID of the offspring.}
\item{ParentId}{Integer, ID of the parent.}
}
}
\source{
BiostatsUSC
}
\description{
A fake dataset containing the parent-offspring relations between genes.
This dataset is inteded for testing only.
}
