% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gain_loss_prob}
\alias{gain_loss_prob}
\title{Gain/Loss probabilities}
\usage{
gain_loss_prob(mu)
}
\arguments{
\item{mu}{A numeric vector of length 2. Gain and Loss probailities.}
}
\value{
a 2x2 matrix with gain loss probabilities, with rows and columns
corresponding to states (0,1) of parent and offspring respectively.
}
\description{
Generates a 2x2 matrix with gain loss probabilities, with rows and columns
corresponding to states (0,1) of parent and offspring respectively.
}

