% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{get_offspring}
\alias{get_offspring}
\alias{as.phylo.phylo_offspring}
\alias{plot.phylo_offspring}
\title{Find the offspring of each node.}
\usage{
get_offspring(data_exper, leafidvar, data_tree, nodeidvar, parentidvar,
  funvars = NULL)

\method{as.phylo}{phylo_offspring}(x, ...)

\method{plot}{phylo_offspring}(x, y = NULL, tip.color = NULL, ...)
}
\arguments{
\item{data_exper}{A data.frame with the experimental data}

\item{leafidvar}{A character scalar with the name of the leaf id variable
in \code{data_exper}.}

\item{data_tree}{A data.frame with the tree.}

\item{nodeidvar}{A character scalar with the name of the node id variable
in \code{data_tree}.}

\item{parentidvar}{A character scalar with the name of the parent id variable
in \code{data_tree}.}

\item{funvars}{A character vector with the names of the function indicator
(0-1) variables in \code{data_exper}. If not provided, it is assumed that
all but \code{leafidvar} are function variables.}

\item{x}{An object of class \code{phylo_offspring}.}

\item{...}{Further arguments passed to the method.}

\item{y}{Ignored.}

\item{tip.color}{See \code{\link[ape:plot.phylo]{plot.phylo}}}
}
\value{
A list of length \eqn{n} with relative position of offspring
of each node, with respect to \code{data_exper}, starting from 0.
}
\description{
Find the offspring of each node.

Plot and print methods for \code{phylo_offspring} objects
}
\details{
Plotting is done via \code{\link[ape:plot.phylo]{plot.phylo}} 
from the \CRANpkg{ape} package.

\code{as.phylo} requires loading the \code{ape} package.
}
\examples{
# Example with lots of data -------------------------------------------------
# Loading data
data(experiment)
data(tree)

ans <- get_offspring(
    experiment, "LeafId", 
    tree, "NodeId", "ParentId"
)

# We can visualize it
plot(ans)

# Example with less data ----------------------------------------------------

ans <- get_offspring(
    fakeexperiment, "LeafId", 
    faketree, "NodeId", "ParentId"
)
 
# We can visualize it
plot(ans)
}
\seealso{
Other Data management functions: \code{\link{as.phylo}},
  \code{\link{recode_vertices}}
}
