% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{internal_prob}
\alias{internal_prob}
\title{Internal node probabilities}
\usage{
internal_prob(Pr, M, S, noffspring, offspring)
}
\arguments{
\item{Pr}{Probabilities (already with leaf probs).}

\item{M}{Gain/Loss probabilities (see equation 4 of math.pdf)}

\item{S}{States}

\item{noffspring}{Number of offspring}

\item{offspring}{List of offspring}
}
\description{
Internal node probabilities
}

