% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leaf_prob}
\alias{leaf_prob}
\title{Leaf probabilities}
\usage{
leaf_prob(Z, S, psi, noffspring)
}
\arguments{
\item{Z}{A matrix of size \eqn{N\times P}{N*P} with values \code{(0,1,9)}.}

\item{S}{A matrix of size \eqn{2^P*P} as returned by \code{\link{states}}.}

\item{psi}{A numeric vector of length 2.}

\item{noffspring}{A numeric vector of length \eqn{N}. Number of offspring
per node.}
}
\description{
Leaf probabilities
}
