% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{leafs}
\alias{leafs}
\alias{leafs.phylo}
\alias{leafs.aphylo}
\title{Extract leaf labels}
\usage{
leafs(x, ...)

\method{leafs}{phylo}(x, ...)

\method{leafs}{aphylo}(x, ...)
}
\arguments{
\item{x}{A phylogenetic tree.}

\item{...}{Ignored.}
}
\value{
\code{leafs} returns a character vector with the names of the leafs
}
\description{
Extract leaf labels
}
\examples{
set.seed(1)
ans <- sim_tree(10)
leafs(ans)
}
