% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{list_offspring}
\alias{list_offspring}
\title{List each nodes' offspring}
\usage{
list_offspring(x)
}
\arguments{
\item{x}{An object of class \code{phylo} or \code{aphylo}.}
}
\value{
List of length \code{n} (total number of nodes).
}
\description{
For each node in a tree, lists all its offspring.
}
\examples{
# A simple example with phylo tree ------------------------------------------

set.seed(4)
x <- ape::rtree(10)
list_offspring(x)

}
