% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.r
\name{mcmc}
\alias{mcmc}
\title{Markov Chain Monte Carlo}
\usage{
mcmc(fun, initial, nbatch = 20000L, scale = 1, burnin = 1000L,
  ub = rep(1 - 1e-20, length(initial)), lb = rep(1e-20, length(initial)))
}
\arguments{
\item{fun}{A function. Returns the log-likelihood}

\item{initial}{A numeric vector. Initial values of the parameters.}

\item{nbatch}{Integer scalar. Number of MCMC runs.}

\item{scale}{Numeric scalar. Step size, \eqn{z\times scale}{z*scale}, where
\eqn{z\sim N(0,1)}{z~N(0,1)}.}

\item{burnin}{Integer scalar. Number of burn-in samples.}
}
\value{
\item{batch}{Numeric matrix of size \code{length(initial) x nbatch}.}
\item{final}{Numeric vector of length \code{length(initial)}. Final state of the parameters.}
\item{initial}{Numeric vector of length \code{length(initial)}. 
Initial state of the parameters.}
\item{fun}{Objective function.}
}
\description{
Markov Chain Monte Carlo
}
\details{
This function implements MCMC using the Hastings ratio with
scaled standard normal propositions for each parameter. For each parameter
the transition function is 

\deqn{
\theta' = \theta + scale*z
}

Where \eqn{z} has standard normal distribution.
}

