% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{mislabel}
\alias{mislabel}
\title{Switch labels acoording to mislabeling probabilities}
\usage{
mislabel(atree, psi)
}
\arguments{
\item{atree}{An object of class \link{aphylo}.}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{mu}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{eta}{Numeric vector of length 2. Annotation bias probabilities (see \code{\link{LogLike}}).}

\item{Pi}{Numeric scalar. Root node probability of having the function (see \code{\link{LogLike}}).}
}
\value{
An object of class \link{aphylo} with modified labels.
}
\description{
Switch labels acoording to mislabeling probabilities

\code{}
}
\examples{

set.seed(131)
x <- sim_annotated_tree(5, P=2, psi=c(0,0))
x$tip.annotation

# Flipping 0s to 1s and vice versa
mislabel(x, psi = c(1,1))$tip.annotation

}
