% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiaphylo.r
\name{multiAphylo}
\alias{multiAphylo}
\alias{c.aphylo}
\alias{c.multiAphylo}
\title{Building Lists of Annotated Trees}
\usage{
\method{c}{aphylo}(...)

\method{c}{multiAphylo}(...)
}
\arguments{
\item{...}{One or several object of class \code{aphylo} or \code{multiAPhylo}.}
}
\description{
This is equivalent to what \link[ape:c.phylo]{ape::c.phylo} does.
}
\examples{
data(fakeexperiment)
data(faketree)
ans <- new_aphylo(fakeexperiment[,2:3], faketree)
c(ans, ans)
}
