% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{phylo_offspring_to_igraph}
\alias{phylo_o_to_phylo}
\alias{phylo_offspring_to_igraph}
\title{Convert \code{phylo_offspring} to \code{igraph} objects.}
\usage{
phylo_offspring_to_igraph(x)

phylo_o_to_phylo(x)
}
\arguments{
\item{x}{An object of class \code{phylo_offspring}}
}
\description{
Convert \code{phylo_offspring} to \code{igraph} objects.
}

