% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{phylo_offsprings_to_igraph}
\alias{phylo_o_to_phylo}
\alias{phylo_offsprings_to_igraph}
\title{Convert \code{phylo_offsprings} to \code{igraph} objects.}
\usage{
phylo_offsprings_to_igraph(x)

phylo_o_to_phylo(x)
}
\arguments{
\item{x}{An object of class \code{phylo_offsprings}}
}
\description{
Convert \code{phylo_offsprings} to \code{igraph} objects.
}

