% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{phylogenetic}
\alias{phylogenetic}
\alias{phylogenetic-package}
\title{Statistical Inference in Annotated Phylogenetic Trees}
\description{
Statistical Inference in Annotated Phylogenetic Trees
}
