% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pies.r
\name{piechart}
\alias{piechart}
\title{A flexible piechart.}
\usage{
piechart(x, add = FALSE, r = 1, doughnut = 0, origin = c(0, 0),
  smooth = 100, labs = NULL, text.off = 0.1, text.args = list(), ...)
}
\arguments{
\item{x}{Numeric vector. Values that specify the area of the slices.}

\item{add}{Logical scalar. When \code{TRUE} it is added to the current device.}

\item{r}{Numeric scalar. Radious of the pie.}

\item{doughnut}{Numeric scalar. Radious of the inner circle (doughnut).}

\item{origin}{Numeric vector of length 2. Coordinates of the origin.}

\item{smooth}{Numeric scalar. Smoothness of the slices curve.}

\item{labs}{Character vector of length \code{length(x)}. Passed to
\code{\link[graphics:text]{text}}.}

\item{text.off}{Numeric scalar. Proportion that the text should be off (on)
the circle with respect to \code{r}.}

\item{text.args}{List. Further arguments passed to \code{\link[graphics:text]{text}}.}

\item{...}{Further arguments passed to \code{\link[graphics:polygon]{polygon}}.}
}
\value{
A list with two elements:
\item{slices}{A list of length \code{length(x)} with the coordinates of each
  slice.}
\item{textcoords}{A numeric matrix of size \code{length(x)*2} with 
  coordinates where the labels can be put at.}
}
\description{
A flexible piechart.
}
\examples{
 
vals <- c(1,2,3,10)
piechart(vals, col=grDevices::blues9[5:8], border=NA, doughnut = .5, r=.75, labs=vals, text.off = .05)
piechart(vals, col=grDevices::blues9[3:6], border=NA, doughnut = .3, r=.5, add=TRUE)
piechart(vals, col=grDevices::blues9[1:4], border=NA, doughnut = .1, r=.3, add=TRUE)
}
