% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{plot.aphylo_prediction_score}
\alias{plot.aphylo_prediction_score}
\alias{plot-prediction}
\title{Visualize predictions}
\usage{
\method{plot}{aphylo_prediction_score}(x, y = NULL,
  main = "Prediction Accuracy: Observed versus predicted values",
  main.colorkey = "Probability of Functional Annotation",
  which.fun = seq_len(ncol(x$expected)), include.labels = NULL,
  labels.col = "black", ...)
}
\arguments{
\item{x}{An object of class \code{aphylo_prediction_score}.}

\item{y}{Ignored.}

\item{main}{Passed to \code{title}.}

\item{main.colorkey}{Character scalar. Title of the colorkey (optional).}

\item{which.fun}{Integer vector. Which function to plot.}

\item{include.labels}{Logical scalar. When \code{TRUE}, draws nice labels
at each slice which by default are specified as the rownames of \code{x$expected}.
This is mostly useful when the number of predictions is small.}

\item{labels.col}{Character scalar. Color of the labels.}

\item{...}{Ignored}
}
\description{
Visualize predictions
}
\details{
If \code{include.labels = NULL} and \code{ncol(x$expected) > 40},
then \code{include.labels=FALSE} by default.
}
