% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loglike.r
\name{plot_logLik}
\alias{plot_logLik}
\alias{plot_logLik.aphylo_estimates}
\title{Plot LogLikelihood function of experimental data}
\usage{
plot_logLik(x, sets, ...)

\method{plot_logLik}{aphylo_estimates}(x, sets, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=aphylo]{aphylo()}}}

\item{sets}{(optional) Character matrix of size \code{2 x # of combinations}.
contains the names of the pairs to plot. If nothing passed, the function will
generate all possible combinations as \code{combn(names(params), 2)}.}

\item{...}{Aditional parameters to be passed to \code{plotfun}.}
}
\description{
Plot LogLikelihood function of experimental data
}
\examples{
# Loading data
data(fakeexperiment)
data(faketree)
O <- new_aphylo(fakeexperiment[,2:3], faketree)

# Nice personalized plot
plot_LogLike(O, nlevels = 60, plotfun = persp, theta = -pi*20, 
  shade=.7, border="darkblue", phi=30, scale=TRUE, 
  par.args = list(mar=c(1, 1, 1, 1), oma=c(0,0,4,0)),
  ticktype = "detailed")
  
# Adding title
mtext(
  "LogLikelihood",
  side=3, outer=FALSE, line = 2, cex=1.25)
}
