% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loglike.r
\name{plot_LogLike}
\alias{plot_LogLike}
\title{Plot LogLikelihood function of experimental data}
\usage{
plot_LogLike(x, psi_range = c(1e-05, 0.1), mu_range = c(1e-05, 0.9999),
  Pi_range = c(1e-05, 0.9999), nlevels = 30, plotfun = persp,
  par.args = list(), ...)
}
\arguments{
\item{x}{An object of class \code{\link[phylogenetic:get_offspring]{phylo_offspring}}}

\item{psi_range}{Numeric vector of length 2. Domain of \eqn{psi}.}

\item{mu_range}{Numeric vector of length 2. Domain of \eqn{mu}.}

\item{Pi_range}{Numeric vector of length 2. Domain of \eqn{pi}.}

\item{nlevels}{Integer scalar. Number of levels of each parameter to create.}

\item{plotfun}{Function. Either \code{\link[graphics:contour]{contour}},
\code{\link[graphics:persp]{persp}}, or other similar function that takes at
least 3 parameters, \code{x,y,z}.}

\item{par.args}{List of arguments to be passed to \code{\link{par}} before
\code{plotfun} is called.}

\item{...}{Aditional parameters to be passed to \code{plotfun}.}
}
\description{
Plot LogLikelihood function of experimental data
}
\examples{
# Loading data
data(experiment)
data(tree)
O <- get_offspring(
  experiment, "LeafId", 
  tree, "NodeId", "ParentId"
)

# Nice personalized plot
plot_LogLike(O, nlevels = 60, plotfun = persp, theta = -pi*20, 
  shade=.7, border="darkblue", phi=30, scale=TRUE, 
  par.args = list(mar=c(1, 1, 1, 1), oma=c(0,0,4,0)),
  ticktype = "detailed")
  
# Adding title
mtext(
  "LogLikelihood",
  side=3, outer=FALSE, line = 2, cex=1.25)
}
