% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-bis.r
\name{posterior-probabilities}
\alias{posterior-probabilities}
\alias{predict_pre_order}
\alias{predict_brute_force}
\title{Posterior probabilities based on parameter estimates}
\usage{
predict_pre_order(atree, psi, mu, eta, Pi, ...)

predict_brute_force(atree, psi, mu, Pi)
}
\arguments{
\item{atree}{A tree of class \link{aphylo}}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{mu}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{eta}{Numeric vector of length 2. Annotation bias probabilities (see \code{\link{LogLike}}).}

\item{Pi}{Numeric scalar. Root node probability of having the function (see \code{\link{LogLike}}).}

\item{...}{Ignored.}
}
\description{
The function \code{predict_pre_order} uses a pre-order algorithm to compute the
posterior probabilities, whereas the \code{predict_brute_force} computes posterior
probabilities generating all possible cases.
}
\details{
The function \code{predict_brute_force} is only intended for testing. For predictions
after estimating the model, see \link{predict.aphylo_estimates}.
}
