% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{prediction_score}
\alias{prediction_score}
\alias{prediction_score.default}
\alias{prediction_score.aphylo_estimates}
\title{Calculate prediction score (quality of prediction)}
\usage{
prediction_score(x, expected, alpha = 0.5, W = NULL)

\method{prediction_score}{default}(x, expected, alpha = 0.5, W = NULL)

\method{prediction_score}{aphylo_estimates}(x, expected,
  alpha = mean(x$dat$tip.annotation == 1L, na.rm = TRUE), W = NULL,
  ...)
}
\arguments{
\item{x}{An object of class \link{aphylo_estimates} or a numeric matrix.}

\item{expected}{Integer vector of length \eqn{n}. Expected values (either 0 or 1).}

\item{alpha}{Numeric scalar. Prior belief of the parameter of the bernoulli distribution
used to compute the random imputation score.}

\item{W}{A square matrix. Must have as many rows as genes in \code{expected}.}

\item{...}{Further arguments passed to \link{predict.aphylo_estimates}}
}
\description{
Calculate prediction score (quality of prediction)
}
\details{
In the case of \code{prediction_score}, \code{...} are passed to
\code{predict.aphylo_estimates}.
}
\examples{
# Example with prediction_score ---------------------------------------------
set.seed(1312)
ap  <- sim_annotated_tree(10, P = 1, Pi=.2, mu=c(.05,.02))
ans <- aphylo_mcmc(
  ap ~ mu + eta + psi + Pi,
  control = list(nsteps=2e4, thin=100),
  priors  = function(x) dbeta(x, 1, 30)
  )
                   
pr <- prediction_score(ans)
with(pr, cbind(Expected = expected, Predicted = predicted))
}
