% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_mat}
\alias{prob_mat}
\title{Probability Matrix}
\usage{
prob_mat(pr)
}
\arguments{
\item{pr}{A numeric vector of length 2.}
}
\value{
a 2x2 matrix with 0/1 probabilities, with rows and columns
corresponding to states (0,1) of parent and offspring respectively.
}
\description{
Generates a 2x2 matrix with probabilities for states 0/1, with rows and columns
corresponding to states (0,1) of parent and offspring respectively.
}

