% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probabilities}
\alias{probabilities}
\title{State probabilities}
\usage{
probabilities(annotations, mu, psi, S, offspring)
}
\arguments{
\item{annotations}{Matrix of size \eqn{n\times P}{n*P}. Annotation data, each column is a functional annotation coded as \code{1} if the node has the function, \code{0} if the node doesn't has the function and \code{9} or \code{NA} otherwise (no information).}

\item{mu}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{S}{Integer matrix of size \eqn{2^P\times P}{2^P * P}. See \code{\link{states}}.}

\item{offspring}{List of length \eqn{n}. Offspring of each node (see \code{\link{new_aphylo}}).}
}
\value{
A numeric matrix of size \eqn{n\times 2^P}{n * 2^P} with state
probabilities for each node.
}
\description{
Compute the state probabilities for each node in the tree using the peeling
algorithm. This function is the horse-power of the function \code{\link{LogLike}}
so it is not intended to be used directly.
}
