% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sifter.R
\name{read_nhx}
\alias{read_nhx}
\title{Read New Hampshire eXtended format for trees}
\usage{
read_nhx(fn)
}
\arguments{
\item{fn}{Full path to the tree file}
}
\value{
A list with the following elements:
\itemize{
\item tree An object of class \code{ape}
\item edge Edge annotations (length and other annotations)
\item nhx A list of annotations NHX
}
}
\description{
Read New Hampshire eXtended format for trees
}
