% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sifter.R
\name{read_pli}
\alias{read_pli}
\title{Read PLI files from SIFTER}
\usage{
read_pli(fn, dropNAs = TRUE)
}
\arguments{
\item{fn}{Full path to the file}

\item{dropNAs}{Logical scalar. When \code{TRUE}, the function will discard any
protein that has no annotations.}
}
\value{
A data table object including the following columns:
\itemize{
\item name: Used to match UniProtKB data and GOA,
\item number,
\item go: A list of the GO annotations
\item moc: Evidence code
\item fam: Name of the family
}
}
\description{
Read PLI files from SIFTER
}
\seealso{
Other reading: 
\code{\link{panther-tree}},
\code{\link{read_nhx}()}
}
\concept{reading}
