% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{recode_vertices}
\alias{recode_vertices}
\title{Relabel a tree using \pkg{ape}'s convention}
\usage{
recode_vertices(offspring, parent, root_label = "0")
}
\arguments{
\item{offspring}{Vector. Labels of offspring.}

\item{parent}{Vector. Labels of parents.}

\item{root_label}{Character scalar. Label of the root node.}
}
\value{
A list with three components:
\item{edges}{Integer Matrix of size \code{length(offspring)*2}. Recoded edgelist}
\item{nodes}{Named character scalar of size \code{n}. Each node's id and label.}
\item{isleaf}{Logical vector of size \code{n}. \code{TRUE} when the node is a leaf node.}
}
\description{
Relabel a tree using \pkg{ape}'s convention
}
\examples{

# A simple example ----------------------------------------------------------
set.seed(1231)

# Generating a random tree: Observe that the nodes are coded such that
# for all (i,j) that are offspring and parent i > j. This is a requirement
# of the peeling algorithm
ans0 <- sim_tree(30)

# In the ape package, nodes are labeled such that leafs are from 1 to n
# root node is n+1 and interior nodes can have whatever label.
ans1 <- recode_vertices(ans0[,1], ans0[,2])

head(ans0)
head(ans1)

}
\seealso{
Other Data management functions: \code{\link{as.phylo}},
  \code{\link{get_offspring}}
}
