% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{root_node_prob}
\alias{root_node_prob}
\title{Root node probabilities}
\usage{
root_node_prob(pi, S)
}
\arguments{
\item{pi}{A numeric vector of length 2 with root node probabilities.}

\item{S}{A numeric matrix of size 2^PxP with all possible states as
generated by \code{\link{states}}

Generates a vector of length 2^P with the root node probabilities
per state}
}
\description{
Root node probabilities
}
