% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{sim_annotated_tree}
\alias{sim_annotated_tree}
\title{Simulation of Annotated Phylogenetic Trees}
\usage{
sim_annotated_tree(n = NULL, tree = NULL, P = 1L, psi = c(0.05, 0.05),
  mu = c(0.1, 0.05), eta = c(1, 1), Pi = 1,
  informative = getOption("aphylo.informative", FALSE), maxtries = 20L)
}
\arguments{
\item{n}{Integer scalar. Number of leafs. If not specified, then}

\item{tree}{An object of class \link[ape:read.tree]{phylo}.}

\item{P}{Integer scalar. Number of functions to generate.}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{mu}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{eta}{Numeric vector of length 2. Annotation bias probabilities (see \code{\link{LogLike}}).}

\item{Pi}{Numeric scalar. Root node probability of having the function (see \code{\link{LogLike}}).}

\item{informative, maxtries}{Passed to \link{sim_fun_on_tree}.}
}
\value{
An object of class \link{aphylo}
}
\description{
Simulation of Annotated Phylogenetic Trees

\code{}
}
\examples{
# A simple example ----------------------------------------------------------

set.seed(1231)
ans <- sim_annotated_tree(n=500)

}
\concept{Simulation Functions}
