% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_phylo}
\alias{sim_phylo}
\title{Simulate functions}
\usage{
sim_phylo(offspring, noffspring, psi, mu, Pi)
}
\arguments{
\item{offspring}{A List of length \eqn{N} with the set of offspring of
each node.}

\item{noffspring}{An integer vector of length \eqn{N} with the number of
offspring per node.}

\item{psi}{A numeric vector of length 2 (see details).}

\item{mu}{A numeric vector of length 2 (see details).}

\item{Pi}{A numeric vector of length 2 (see details).}
}
\value{
An integer matrix with values 9, 0 and 1.
}
\description{
Simulate functions
}
\examples{
# Example 1 ----------------------------------------------------------------
# Loading the data
data(experiment)
data(tree)
  
# Preprocessing the data
O <- get_offspring(experiment, "LeafId", tree, "NodeId", "ParentId")
    
# Simulating
ans <-
  with(O, sim_phylo(
      offspring,
      noffspring,
      psi = c(.001, .05) * 0,
      mu = c(.01, .05),
      Pi = c(.5, .5)
  ))
      
# Tabulating results
table(ans)


}
