% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{states}
\alias{states}
\title{Matrix of states}
\usage{
states(P)
}
\arguments{
\item{P}{Integer scalar. Number of functions.}
}
\value{
A matrix of size 2^P by P with all the possible
(0,1) combinations of functions.
}
\description{
Matrix of states
}
\examples{
states(3)
}
