% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.r
\name{tree}
\alias{tree}
\title{Phylogenetic Tree}
\format{A data frame with 326 rows and 3 variables:
\describe{
  \item{NodeId}{Integer, ID of the offspring.}
  \item{TypeId}{Integer, Type of node (unusued).}
  \item{ParentId}{Integer, ID of the parent.}
}}
\source{
BiostatsUSC
}
\description{
A dataset containing the parent-offspring relations between genes.
}

