% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sifter.R
\name{write_pli}
\alias{write_pli}
\title{Write pli files used by SIFTER}
\usage{
write_pli(family_id, protein_name, protein_number, go_number,
  moc = "EXP", file = "")
}
\arguments{
\item{family_id}{Character scalar. Name of the family}

\item{protein_name, protein_number, go_number, moc}{Vectors of the same length}
}
\description{
Write pli files used by SIFTER
}
