% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/aphylo-ape-info.r
\name{Ntip.aphylo_pruner}
\alias{Ntip.aphylo_pruner}
\alias{Nannotated.aphylo_pruner}
\alias{Nann.aphylo_pruner}
\alias{ape-methods}
\alias{Nedge.aphylo}
\alias{Nedge.aphylo_estimates}
\alias{Nedge.multiAphylo}
\alias{Nnode.aphylo}
\alias{Nnode.aphylo_estimates}
\alias{Nnode.multiAphylo}
\alias{Nnode.aphylo_pruner}
\alias{Ntip.aphylo}
\alias{Ntip.aphylo_estimates}
\alias{Ntip.multiAphylo}
\title{APE wrappers}
\usage{
\method{Ntip}{aphylo_pruner}(phy)

\method{Nannotated}{aphylo_pruner}(phy)

\method{Nann}{aphylo_pruner}(phy)

\method{Nedge}{aphylo}(phy)

\method{Nedge}{aphylo_estimates}(phy)

\method{Nedge}{multiAphylo}(phy)

\method{Nnode}{aphylo}(phy, ...)

\method{Nnode}{aphylo_estimates}(phy, ...)

\method{Nnode}{multiAphylo}(phy, ...)

\method{Nnode}{aphylo_pruner}(phy, ...)

\method{Ntip}{aphylo}(phy)

\method{Ntip}{aphylo_estimates}(phy)

\method{Ntip}{multiAphylo}(phy)
}
\arguments{
\item{phy}{An object of class \link{aphylo} (see \link[ape:Nnode]{ape::Nnode}).}

\item{...}{Further arguments passed to the method.}
}
\value{
Integer with the number of edges, nodes, or tips accordignly.
}
\description{
APE wrappers
}
\details{
Functions \code{Nedge}, \code{Nnode}, and \code{Ntip} have S3 methods in the \code{ape} package.
}
\seealso{
Other information: \code{\link{aphylo-info}}
}
\concept{information}
