% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo-ape-info.R
\name{ape-methods}
\alias{ape-methods}
\title{Available methods from the APE package}
\value{
Integer with the number of edges, nodes, or tips accordignly.
}
\description{
The generics \code{\link[ape:Nedge]{ape::Nedge()}}, \code{\link[ape:Nnode]{ape::Nnode()}}, and \code{\link[ape:Ntip]{ape::Ntip()}} can be used
directly on objects of class \link{aphylo}, \link{aphylo_estimates}, \link{multiAphylo}
}
\seealso{
Other information: \code{\link{aphylo-info}}
}
\concept{information}
