% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.r
\name{ape-methods}
\alias{ape-methods}
\alias{Nedge.aphylo}
\alias{Nnode.aphylo}
\alias{Ntip.aphylo}
\alias{Nann}
\alias{Nann.aphylo}
\alias{Nann.phylo}
\alias{Nann.multiPhylo}
\title{APE wrappers}
\usage{
\method{Nedge}{aphylo}(phy)

\method{Nnode}{aphylo}(phy, ...)

\method{Ntip}{aphylo}(phy)

Nann(phy)

\method{Nann}{aphylo}(phy)

\method{Nann}{phylo}(phy)

\method{Nann}{multiPhylo}(phy)
}
\arguments{
\item{phy}{An object of class \link{aphylo} (see \link[ape:Nnode]{ape::Nnode}).}

\item{...}{Further arguments passed to the method.

Functions \code{Nedge}, \code{Nnode}, and \code{Ntip} have S3 methods in the \code{ape} package.

The function \code{Nann}, which reports the number of annotations that the tree has,
is not a function present in ape, but has a method for
both \code{phylo} and \code{multiPhylo} objects which returns a warning an 0L.}
}
\value{
Integer with the number of edges, nodes, or tips accordignly.
}
\description{
APE wrappers
}
