% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{aphylo-class}
\alias{aphylo-class}
\alias{new_aphylo}
\title{Annotated Phylogenetic Tree}
\usage{
new_aphylo(annotations, edges)
}
\arguments{
\item{annotations}{Matrix of size \eqn{n\times P}{n*P}. Annotation data, one column per function coded as \code{1} if the node has the function, \code{0} if the node doesn't has the function and \code{9} otherwise (no information).}

\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}

\item{leafidvar}{A character scalar with the name of the leaf id variable
in \code{annotations}.}

\item{funvars}{A character vector with the names of the function indicator
(0-1) variables in \code{data_exper}. If not provided, it is assumed that
all but \code{leafidvar} are function variables.}
}
\value{
A list of length \eqn{n} with relative position of offspring
of each node, with respect to \code{data_exper}, starting from 0.
}
\description{
Annotated Phylogenetic Tree

\code{}
}
\details{
Plotting is done via \code{\link[ape:plot.phylo]{plot.phylo}} 
from the \CRANpkg{ape} package.
}
\examples{
# A simple example ----------------------------------------------------------

data(fakeexperiment)
data(faketree)
ans <- new_aphylo(fakeexperiment, faketree, "LeafId")
 
# We can visualize it
plot(ans)
}
\seealso{
Other Data management functions: \code{\link{as.apephylo}},
  \code{\link{as_po_tree}}
}
