% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{aphylo-class}
\alias{aphylo-class}
\alias{new_aphylo}
\title{Annotated Phylogenetic Tree}
\usage{
new_aphylo(tip.annotation, tree, node.annotation = NULL)
}
\arguments{
\item{tip.annotation, node.annotation}{Annotation data. See \linkS4class{aphylo}.}

\item{tree}{An object of class \link[ape:read.tree]{phylo}}

\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}
}
\value{
A list of class \code{aphylo} with the following elements:
\item{tree}{An object of class \link[ape:read.tree]{phylo}.}
\item{tip.annotation}{An integer matrix. Tip (leaf) nodes annotations.}
\item{node.annotation}{An integer matrix (optional). Internal nodes
annotations.}
\item{offspring}{A list. List of offspring of each node.}
\item{pseq}{Integer vector. The pruning sequence (postorder).}
}
\description{
The \code{aphylo} class tree holds both the tree structure represented as a
partially ordered phylogenetic tree, and node annotations. While annotations
are included for both leafs and inner nodes, the algorithms included in this
package only uses the leaf annotations.
}
\details{
Plotting is done via \code{\link[ggtree:ggtree]{ggtree::ggtree()}}
from the \pkg{ggtree} package (Bioconductor).
}
\examples{
# A simple example ----------------------------------------------------------

data(fakeexperiment)
data(faketree)
ans <- new_aphylo(fakeexperiment[,2:3], faketree)
 
# We can visualize it
plot(ans)
}
\seealso{
Other aphylo methods: \code{\link{aphylo-methods}}
}
\concept{Data management functions}
\concept{aphylo methods}
