% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.r
\name{aphylo-index}
\alias{aphylo-index}
\alias{[.aphylo}
\alias{[<-.aphylo}
\title{Indexing aphylo objects}
\usage{
\method{[}{aphylo}(x, i, j, drop = FALSE)

\method{[}{aphylo}(x, i, j) <- value
}
\arguments{
\item{x}{An object of class \link{aphylo}.}

\item{i, j}{Integer vector. Indices of genes or functions.}

\item{drop}{Logical scalar. When \code{TRUE}, the function returns a matrix of
annotations. Otherwise an object of class \code{aphylo}.}

\item{value}{Integer vector. Replacing values, can be either \code{c(0, 1, 9, NA)}.}
}
\description{
Indexing aphylo objects
}
\details{
The subsetting method allows selecting one or more annotations from
the \link{aphylo} object. Whenever \code{i} is specified, then aphylo returns the corresponding
annotations.
}
