% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo-ape-info.r
\name{aphylo-info}
\alias{aphylo-info}
\alias{Nann}
\alias{Nann.aphylo}
\alias{Nann.multiAphylo}
\alias{Nann.aphylo_estimates}
\alias{Nannotated}
\alias{Nannotated.aphylo}
\alias{Nannotated.multiAphylo}
\alias{Nannotated.aphylo_estimates}
\alias{Ntrees}
\alias{Ntrees.aphylo}
\alias{Ntrees.multiAphylo}
\alias{Ntrees.aphylo_estimates}
\title{Information about \code{aphylo} and \code{multiAphylo} objects}
\usage{
Nann(phy)

\method{Nann}{aphylo}(phy)

\method{Nann}{multiAphylo}(phy)

\method{Nann}{aphylo_estimates}(phy)

Nannotated(phy)

\method{Nannotated}{aphylo}(phy)

\method{Nannotated}{multiAphylo}(phy)

\method{Nannotated}{aphylo_estimates}(phy)

Ntrees(phy)

\method{Ntrees}{aphylo}(phy)

\method{Ntrees}{multiAphylo}(phy)

\method{Ntrees}{aphylo_estimates}(phy)
}
\arguments{
\item{phy}{Either an object of class \link{aphylo}, \link{multiAphylo}, or
\link{aphylo_estimates}.}
}
\value{
If \code{phy} is of class \code{aphylo}, then a single scalar.
otherwise, if \code{phy} is of class \code{multiAphylo}
}
\description{
Information about annotations, in particular, number of annotations (\code{Nann}),
number of annotated leaves (\code{Nannotated}), number of unnanotated leaves
(\code{Nunannotated}), and number of trees (\code{Ntrees}).
}
\seealso{
Other information: \code{\link{ape-methods}}
}
\concept{information}
