% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{aphylo-methods}
\alias{aphylo-methods}
\alias{plot.aphylo}
\alias{print.aphylo}
\alias{summary.aphylo}
\title{Plot and print methods for \code{aphylo} objects}
\usage{
\method{plot}{aphylo}(x, y = NULL, prop = 0.15,
  node.type.col = c(dupl = "red", other = "blue"),
  node.type.size = c(dupl = 0, other = 0), rect.args = list(), ...)

\method{print}{aphylo}(x, ...)

\method{summary}{aphylo}(object, ...)
}
\arguments{
\item{x}{An object of class \code{aphylo}.}

\item{y}{Ignored.}

\item{prop}{Numeric scalar between 0 and 1. Proportion of the device that the
annotations use in \code{plot.aphylo}.}

\item{node.type.col, node.type.size}{Vectors of length 2. In the case of
\code{node.type.col} the color of the duplication and other nodes. \code{node.type.size}
sets the size of circles.}

\item{rect.args}{List of arguments passed to \link[graphics:rect]{graphics::rect}.}

\item{...}{Further arguments passed to \link[ape:plot.phylo]{ape::plot.phylo}.}

\item{object}{An object of class \code{aphylo}.}

\item{loo}{Logical scalar. When \code{loo = TRUE}, predictions are preformed
similar to what a leave-one-out cross-validation scheme would be done
(see \link{predict.aphylo_estimates}).}
}
\value{
In the case of \code{plot.aphylo}, \code{NULL}.

In the case of \code{print.aphylo}, the same \code{aphylo} object
invisible.
}
\description{
Plot and print methods for \code{aphylo} objects
}
\details{
The \code{plot.aphylo} function is a wrapper of \link[ape:plot.phylo]{ape::plot.phylo}.
}
\seealso{
Other aphylo methods: \code{\link{aphylo-class}}
}
\concept{aphylo methods}
