% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{aphylo-methods}
\alias{aphylo-methods}
\alias{plot.aphylo}
\alias{leafs}
\alias{print.aphylo}
\alias{summary.aphylo}
\title{Plot and print methods for \code{aphylo} objects}
\usage{
\method{plot}{aphylo}(x, y = NULL, geom.tiplab.args = list(),
  gheatmap.args = list(), scale.fill.args = list(), ...)

leafs(...)

\method{print}{aphylo}(x, ...)

\method{summary}{aphylo}(object, ...)
}
\arguments{
\item{x}{An object of class \code{aphylo}.}

\item{y}{Ignored.}

\item{geom.tiplab.args}{Further arguments passed to \code{\link[ggtree:ggtree]{ggtree}}}

\item{gheatmap.args}{Further arguments passed to \code{\link[ggtree:ggtree]{ggtree}}}

\item{scale.fill.args}{Further arguments passed to \code{\link[ggtree:ggtree]{ggtree}}}

\item{...}{Further arguments passed to the method.}

\item{object}{An object of class \code{aphylo}.}
}
\value{
In the case of \code{plot.aphylo}, an object of class \code{c("ggtree", "gg", "ggplot")}

\code{leafs} returns a character vector with the names of the leafs

In the case of \code{print.aphylo}, the same \code{aphylo} object
invisible.
}
\description{
Plot and print methods for \code{aphylo} objects
}
\details{
The \code{plot.aphylo} function is a wrapper of \code{\link[ggtree:ggtree]{ggtree}}
that creates a visualization as follows:
\enumerate{
  \item Retrieve the annotations from the \code{\link[=aphylo-class]{aphylo}} object
  \item Create a \code{ggtree} map adding \code{\link[ggtree:geom_tiplab]{geom_tiplab}}s
  \item Use \code{\link[ggtree:gheatmap]{gheatmap}} to add a heatmap.
  \item Set the colors using \code{\link[ggplot2:scale_fill_manual]{scale_fill_manual}}
}
}
\seealso{
Other aphylo methods: \code{\link{aphylo-class}}
}
