% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphylo_hier.R
\name{aphylo_hier}
\alias{aphylo_hier}
\title{Fit a Hierarchical aphylo model}
\usage{
aphylo_hier(
  model,
  params,
  classes,
  ...,
  params0 = NULL,
  hyper_params = NULL,
  env = parent.frame(),
  verbose = TRUE,
  multicore = FALSE,
  nchains = 1L
)
}
\arguments{
\item{model}{A model as specified in \link{aphylo-model}.}

\item{params}{A vector of length 7 with initial parameters. In particular
\code{psi[1]}, \code{psi[2]}, \code{mu[1]}, \code{mu[2]}, \code{eta[1]}, \code{eta[2]} and \code{Pi}.}

\item{params0}{Starting parameters.}

\item{env}{Environment where to evaluate \code{model}.}

\item{priors}{A function to be used as prior for the model (see \link{bprior}).}

\item{check_informative}{Logical scalar. When \code{TRUE} the algorithm
stops with an error when the annotations are uninformative (either 0s or 1s).}

\item{reduced_pseq}{Logical. When \code{TRUE} it will use a reduced peeling sequence
in which it drops unannotated leafs. If the model includes \code{eta} this is set
to \code{FALSE}.}
}
\value{
An object of class \link{aphylo_estimates}.
}
\description{
Fit a Hierarchical aphylo model
}
\seealso{
Other parameter estimation: 
\code{\link{aphylo_mcmc}()},
\code{\link{aphylo_mle}()}
}
\concept{parameter estimation}
