% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approx_geodesic}
\alias{approx_geodesic}
\title{Approximation of Geodesic distances using Matrix Powers}
\usage{
approx_geodesic(edges, nsteps = 5000L, undirected = TRUE, warn = FALSE)
}
\arguments{
\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}

\item{nsteps}{Integer scalar. Number of maximum steps for the approximation.}

\item{undirected}{Logical scalar. When \code{TRUE} (default), the edgelist is treated
as undirected (see details).}

\item{warn}{Logical scalar. When \code{TRUE} shows a warning after no further
steps are needed.}
}
\value{
A square matrix of size \eqn{n} with the shortest path between each
pair of nodes.
}
\description{
Given an adjacency matrix \eqn{A}, the geodesic can be approximated using
its powers, since each \eqn{(i,j)} element of \eqn{A^p} corresponds to the
number of \eqn{p} length steps between nodes \eqn{i} and \eqn{j}.
}
\details{
When \code{undirected = TRUE}, the function extends \code{edges} such that
\code{edges = rbind(edges, edges[,2:1])}.
}
\references{
This is a modified version of the function of the same name in the
R package \CRANpkg{netdiffuseR}.
}
\author{
George G. Vega Yon
}
