% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_po_tree}
\alias{as_po_tree}
\title{Recodes an edgelist as a Partially Ordered Tree}
\usage{
as_po_tree(edges)
}
\arguments{
\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}
}
\description{
Recodes an edgelist as a Partially Ordered Tree
}
