% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r, R/methods.r
\name{get_offspring}
\alias{get_offspring}
\alias{plot.phylo_offspring}
\title{Find the offspring of each node.}
\usage{
get_offspring(data_exper, leafidvar, data_tree, nodeidvar, parentidvar,
  funvars = NULL)

\method{plot}{phylo_offspring}(x, y = NULL, layout = NULL, ...)
}
\arguments{
\item{data_exper}{A data.frame with the experimental data}

\item{leafidvar}{A character scalar with the name of the leaf id variable
in \code{data_exper}.}

\item{data_tree}{A data.frame with the tree.}

\item{nodeidvar}{A character scalar with the name of the node id variable
in \code{data_tree}.}

\item{parentidvar}{A character scalar with the name of the parent id variable
in \code{data_tree}.}

\item{funvars}{A character vector with the names of the function indicator
(0-1) variables in \code{data_exper}. If not provided, it is assumed that
all but \code{leafidvar} are function variables.}

\item{x}{An object of class \code{phylo_offspring}.}

\item{y}{Ignored.}

\item{layout}{Passed to \code{\link[igraph:plot.igraph]{igraph:plot.igraph}}}

\item{...}{Further arguments passed to the method.}
}
\value{
A list of length \eqn{n} with relative position of offspring
of each node, with respect to \code{data_exper}, starting from 0.
}
\description{
Find the offspring of each node.

Plot and print methods for \code{phylo_offspring} objects
}
\details{
In the case of the plot method, by defult layout is set to
be \code{\link[igraph:layout_with_sugiyama]{layout_with_sugiyama}}
with parameter \code{maxiter=200}.
}
\examples{
# Loading data
data(experiment)
data(tree)

ans <- get_offspring(
    experiment, "LeafId", 
    tree, "NodeId", "ParentId"
)

str(ans)

# We can visualize it
plot(ans, vertex.size=5, vertex.label=NA)
}

