% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{internal_prob}
\alias{internal_prob}
\title{Internal node probabilities}
\usage{
internal_prob(Pr, M, S, noffsprings, offsprings)
}
\arguments{
\item{Pr}{Probabilities (already with leaf probs).}

\item{M}{Gain/Loss probabilities (see equation 4 of math.pdf)}

\item{S}{States}

\item{noffsprings}{Number of offsprings}

\item{offsprings}{List of offsprings}

\item{pi}{Root node state probabilities (as returned by root_node_prob)}
}
\description{
Internal node probabilities
}

