% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{list_offspring}
\alias{list_offspring}
\title{List each nodes' offspring}
\usage{
list_offspring(x)
}
\arguments{
\item{x}{Either a matrix or an object of class \code{phylo}}
}
\value{
A named list of length \code{n} (total number of nodes). The names of the
list correspond to the ids of the nodes in \code{x}.
}
\description{
For each node in a tree, lists all its offspring.
}
\examples{
# A simple example with phylo tree ------------------------------------------

set.seed(4)
x <- ape::rtree(10)
list_offspring(x)

}
