% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_offspring.R
\name{list_offspring}
\alias{list_offspring}
\alias{list_parents}
\title{List each nodes' offspring or parent}
\usage{
list_offspring(x)

list_parents(x)
}
\arguments{
\item{x}{An object of class \code{phylo} or \code{aphylo}.}
}
\value{
List of length \code{n} (total number of nodes).
}
\description{
For each node in a tree, the functions \code{list_offspring} and \code{list_parents}
lists all its offspring and parents, respectively.
}
\examples{
# A simple example with phylo tree ------------------------------------------

set.seed(4)
x <- ape::rtree(10)
list_offspring(x)

}
