% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLike}
\alias{LogLike}
\title{Computes Log-likelihood}
\usage{
LogLike(Z, offspring, noffspring, psi, mu, Pi, verb_ans = FALSE)
}
\arguments{
\item{Z}{Matrix of size \eqn{n\times P}{n*P}. Annotation data, one column per function coded as \code{1} if the node has the function, \code{0} if the node doesn't has the function and \code{9} otherwise (no information).}

\item{offspring}{List of length \eqn{n}. Offspring of each node (see \code{\link{new_aphylo}}).}

\item{noffspring}{Integer vector of length \eqn{n}. Number of offspring each node has (see \code{\link{new_aphylo}}).}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{mu}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{Pi}{Numeric vector of length 2. Root node probabilities (see \code{\link{LogLike}}).}

\item{verb_ans}{Logical scalar. When \code{FALSE} (default) the function
returns a list with a single scalar (the log-likelihood).}
}
\value{
A list of class \code{phylo_LogLik} with the following elements:
\item{S}{An integer matrix of size \eqn{2^p\times p}{2^p * p} as returned
by \code{\link{states}}.}
\item{Pr}{A numeric matrix of size \eqn{G\times 2^p}{G * 2^p} with node/state
probabilities.}
\item{ll}{A numeric scalar with the log-likelihood value given the chosen
parameters.}
}
\description{
This function computes the log-likelihood of the chosen parameters given
a particular dataset. The arguments \code{Z}, \code{offspring}, and
\code{noffspring} should be as those returned by \code{\link{new_aphylo}}.
For complete Maximum Likelihood Estimation see \code{\link{mle}}.
}
\details{
The parameters to estimate are described as follows:
\enumerate{
\item{\code{psi}: A vector of length 2 with \eqn{\psi_0}{psi[0]} and
\eqn{\psi_1}{psi[1]}, which are the misclassification probabilities fo
\eqn{s_p=0}{s[p]=0} and \eqn{s_p=1}{s[p]=1}
respectively.}
\item{\code{mu}: A vector of length 2 with \eqn{\mu_0}{mu[0]} and
\eqn{\mu_1}{mu[1]} which are the gain and loss probabilities respectively.}
\item{\code{Pi}: A vector of length 2 with \eqn{\pi_0}{pi[0]}} and
\eqn{\pi_1}{pi[1]}, which for now is specified as \eqn{(1 - \pi_1)}{(1 - pi[0]),
which holds the root node probabilities.}
}
}
