% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mislabel}
\alias{mislabel}
\title{Switch labels acoording to mislabeling probabilities}
\usage{
mislabel(atree, psi)
}
\arguments{
\item{atree}{An object of class \link{aphylo}.}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}
}
\value{
An object of class \link{aphylo} with modified labels.
}
\description{
Switch labels acoording to mislabeling probabilities
}
\examples{
set.seed(131)
x <- raphylo(5, P=2, psi=c(0,0))
x$tip.annotation

# Flipping 0s to 1s and vice versa
mislabel(x, psi = c(1,1))$tip.annotation

}
