% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.r
\name{mle}
\alias{mle}
\alias{plot.phylo_mle}
\title{Maximum Likelihood Estimation of the Phylogenetic Tree}
\usage{
mle(params, dat, maxiter = 20L, criter = 1e-05, useABC = FALSE,
  priors = NULL, abcoptim.args = list(maxCycle = 500L, criter = 50L),
  fix.params = c(psi0 = FALSE, psi1 = FALSE, mu0 = FALSE, mu1 = FALSE, Pi =
  FALSE), method.args = list(d = 1e-04), solve.tol = 1e-40)

\method{plot}{phylo_mle}(x, y = NULL, main = "", xlab = "Step",
  ylab = "Log-Likelihood", type = "l", addlegend = TRUE, ...)
}
\arguments{
\item{params}{A vector of length 5 with initial parameters. In particular
\code{psi[1]}, \code{psi[2]}, \code{mu[1]}, \code{mu[2]}, and \code{Pi}.}

\item{dat}{An object of class \code{phylo_offspring} as returned by
\code{\link{get_offspring}}.}

\item{maxiter}{Integer scalar. Maximum number of steps in the Newton-Raphson
algorithm.}

\item{criter}{Numeric scalar. Stoping criteria for the Newton-Raphson algorithm.}

\item{useABC}{Logical scalar. When \code{TRUE}, uses Artificial Bee Colony
optimization algorithm instead.}

\item{priors}{A list of length 3 with functions named \code{psi}, \code{mu},
\code{Pi}}

\item{abcoptim.args}{A list of arguments to be passed to
\code{\link[ABCoptim:abc_cpp]{abc_cpp}} in the \pkg{ABCoptim} package.}

\item{fix.params}{A Logical vector of length 5. Whether or not to fix
a particular parameter to that of what was specified in \code{params}.}

\item{method.args}{A list of arguments passed to \code{\link[numDeriv:jacobian]{hessian,jacobian}}
from the \CRANpkg{numDeriv} package.}

\item{solve.tol}{Numeric scalar passed to \code{\link{solve}}.}

\item{x}{An object of class \code{phylo_mle}.}

\item{y}{Ignored.}

\item{main}{Passed to plot.}

\item{xlab}{Passed to plot.}

\item{ylab}{Passed to plot.}

\item{type}{Passed to plot.}

\item{addlegend}{Logical scalar. When \code{TRUE} adds extra info to the
plot as a legend including ll at the optimum and parameter values.}

\item{...}{Further arguments passed to plot}
}
\value{
A list of class \code{phylo_mle} with the following elements:
\item{par}{A numeric vector of length 5 with the solution.}
\item{hist}{A numeric matrix of size \code{counts*5} with the solution path.}
\item{value}{A numeric scalar with the value of \code{fun(par)}}
\item{fun}{A function (the objective function).}
\item{dat}{The data \code{dat} provided to the function.}
}
\description{
The optimization is done either via \code{abc_cpp} or Newton-Raphson.
}
\examples{

\dontrun{
# Using the package data ----------------------------------------------------
# Loading data
data(experiment)
data(tree)

# Preprocessing the data
O <- get_offspring(
  experiment, "LeafId", 
  tree, "NodeId", "ParentId"
)

# Computing Estimating the parameters 
ans_nr  <- mle(rep(.5,5), O)
ans_abc <- mle(rep(.5,5), O, useABC = TRUE)

# Plotting the path
with(ans_nr, plot(
  - apply(hist, 1, fun),
  type = "l",
  xlab = "Step",
  ylab = "Log-Likelihood"
))

# Computing Estimating the parameters Using Priors for PSI 
mypriors <- function(params) {
    dbeta(params[1:2], 2, 10)
}
ans_nr_dbeta <- mle(rep(.5,5), O, priors = mypriors)
ans_abc_dbeta <- mle(rep(.5,5), O, priors = mypriors, useABC = TRUE)

# Plotting the path
oldpar <- par(no.readonly = TRUE)
par(mfrow = c(2, 2))

plot(ans_nr, main = "No priors NR")
plot(ans_abc, main = "No priors ABC")
plot(ans_nr_dbeta, main = "NR w/ Prior for Psi ~ beta(2,10)")
plot(ans_abc_dbeta, main = "ABC w/ Prior for Psi ~ beta(2,10)")

par(oldpar)
}
# Adding some zeros to the data ---------------------------------------------
set.seed(1231)
mysample <- which(rowSums(experiment[,1:3]) == 27)
mysample <- sample(mysample, 20)
experiment[mysample,1:3] <- 0

O <- get_offspring(
  experiment, "LeafId", 
  tree, "NodeId", "ParentId"
)

ans_nr  <- mle(rep(.5,5), O)
ans_abc <- mle(rep(.5,5), O, useABC = TRUE)


# Computing Estimating the parameters Using Priors for PSI ------------------
mypriors <- function(params) {
  dbeta(params[1:2], 2, 10)
}
ans_nr_dbeta <- mle(rep(.5,5), O, priors = mypriors)
ans_abc_dbeta <- mle(rep(.5,5), O, priors = mypriors, useABC = TRUE)

# Plotting the path
oldpar <- par(no.readonly = TRUE)
par(mfrow = c(2, 2))

plot(ans_nr, main = "No priors NR")
plot(ans_abc, main = "No priors ABC")
plot(ans_nr_dbeta, main = "NR w/ Prior for Psi ~ beta(2,10)")
plot(ans_abc_dbeta, main = "ABC w/ Prior for Psi ~ beta(2,10)")

par(oldpar)

}

