% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.r
\name{mle}
\alias{mle}
\alias{plot.phylo_mle}
\title{Maximum Likelihood Estimation of the Phylogenetic Tree}
\usage{
mle(params, dat, maxiter = 20L, criter = 1e-05, useABC = FALSE,
  priors = NULL, abcoptim.args = list(ub = 1, lb = 0, maxCycle = 500L,
  criter = 50L))

\method{plot}{phylo_mle}(x, y = NULL, main = "", xlab = "Step",
  ylab = "Log-Likelihood", type = "l", addlegend = TRUE, ...)
}
\arguments{
\item{params}{A vector of length 5 with initial parameters. In particular
\code{psi[1]}, \code{psi[2]}, \code{mu[1]}, \code{mu[2]}, and \code{Pi}.}

\item{dat}{An object of class \code{phylo_offspring} as returned by
\code{\link{get_offspring}}.}

\item{maxiter}{Integer scalar. Maximum number of steps in the Newton-Raphson
algorithm.}

\item{criter}{Numeric scalar. Stoping criteria for the Newton-Raphson algorithm.}

\item{useABC}{Logical scalar. When \code{TRUE}, uses Artificial Bee Colony
optimization algorithm instead.}

\item{priors}{A list of length 3 with functions named \code{psi}, \code{mu},
\code{Pi}}

\item{abcoptim.args}{A list of arguments to be passed to
\code{\link[ABCoptim:abc_cpp]{abc_cpp}} in the \pkg{ABCoptim} package.}

\item{x}{An object of class \code{phylo_mle}.}

\item{y}{Ignored.}

\item{main}{Passed to plot.}

\item{xlab}{Passed to plot.}

\item{ylab}{Passed to plot.}

\item{type}{Passed to plot.}

\item{addlegend}{Logical scalar. When \code{TRUE} adds extra info to the
plot as a legend including ll at the optimum and parameter values.}

\item{...}{Further arguments passed to plot}
}
\value{
A list of class \code{phylo_mle} with the following elements:
\item{par}{A numeric vector of length 5 with the solution.}
\item{hist}{A numeric matrix of size \code{counts*5} with the solution path.}
\item{value}{A numeric scalar with the value of \code{fun(par)}}
\item{fun}{A function (the objective function).}
\item{dat}{The data \code{dat} provided to the function.}
}
\description{
The optimization is done either via \code{abc_cpp} or Newton-Raphson.
}
\examples{

\dontrun{
# Loading data
data(experiment)
data(tree)

# Preprocessing the data
O <- get_offspring(
  experiment, "LeafId", 
  tree, "NodeId", "ParentId"
)

# Computing Estimating the parameters ---------------------------------------
ans_nr  <- mle(rep(.5,5), O)
ans_abc <- mle(rep(.5,5), O, useABC = TRUE)

# Plotting the path
with(ans_nr, plot(
  - apply(hist, 1, fun),
  type = "l",
  xlab = "Step",
  ylab = "Log-Likelihood"
))

# Computing Estimating the parameters Using Priors for PSI ------------------
mypriors <- function(params) {
    dbeta(params[1:2], 1, 9)
}
ans_nr_dbeta <- mle(rep(.5,5), O, priors = mypriors)

# Plotting the path
oldpar <- par(no.readonly = TRUE)
par(mfrow = c(2, 2))

plot(ans_nr, main = "No priors NR")
plot(ans_abc, main = "No priors ABC")
plot(ans_nr_dbeta, main = "Prior for Psi ~ beta(1,9)")

par(oldpar)
}
}

