% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiaphylo.R
\name{multiAphylo}
\alias{multiAphylo}
\alias{c.aphylo}
\alias{print.multiAphylo}
\title{Building Lists of Annotated Trees}
\usage{
\method{c}{aphylo}(...)

\method{print}{multiAphylo}(x, ...)
}
\arguments{
\item{...}{One or several object of class \code{aphylo} or \code{multiAPhylo}. Ignored
in the case of \code{print.multiAphylo}.}

\item{x}{An object of class \code{multiAphylo}}
}
\value{
A list of class \code{multiAphylo}. Each element corresponds to a single \code{aphylo}
object.
}
\description{
This is equivalent to what \link[ape:c.phylo]{ape::c.phylo} does.
}
\examples{
data(fakeexperiment)
data(faketree)
ans <- new_aphylo(fakeexperiment[,2:3], tree = as.phylo(faketree))
c(ans, ans)
}
