% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{plot.aphylo}
\alias{plot.aphylo}
\title{Plot and print methods for \code{aphylo} objects}
\usage{
\method{plot}{aphylo}(x, y = NULL, tip.color = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{aphylo}.}

\item{y}{Ignored.}

\item{tip.color}{See \code{\link[ape:plot.phylo]{plot.phylo}}}

\item{...}{Further arguments passed to the method.}
}
\description{
Plot and print methods for \code{aphylo} objects
}
