% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.r
\name{new_aphylo}
\alias{new_aphylo}
\alias{as.phylo.aphylo}
\alias{plot.aphylo}
\title{Annotated Phylogenetic Tree}
\usage{
new_aphylo(annotations, edges, leafidvar, funvars = NULL)

\method{as.phylo}{aphylo}(x, ...)

\method{plot}{aphylo}(x, y = NULL, tip.color = NULL, ...)
}
\arguments{
\item{edges}{Integer matrix with two columns. An edgelist where each row is in the form of \code{(parent, offspring)}.}

\item{leafidvar}{A character scalar with the name of the leaf id variable
in \code{data_exper}.}

\item{funvars}{A character vector with the names of the function indicator
(0-1) variables in \code{data_exper}. If not provided, it is assumed that
all but \code{leafidvar} are function variables.}

\item{x}{An object of class \code{aphylo}.}

\item{...}{Further arguments passed to the method.}

\item{y}{Ignored.}

\item{tip.color}{See \code{\link[ape:plot.phylo]{plot.phylo}}}

\item{data_exper}{A data.frame with the experimental data}

\item{nodeidvar}{A character scalar with the name of the node id variable
in \code{data_tree}.}

\item{parentidvar}{A character scalar with the name of the parent id variable
in \code{data_tree}.}
}
\value{
A list of length \eqn{n} with relative position of offspring
of each node, with respect to \code{data_exper}, starting from 0.
}
\description{
Annotated Phylogenetic Tree

Plot and print methods for \code{aphylo} objects
}
\details{
Plotting is done via \code{\link[ape:plot.phylo]{plot.phylo}} 
from the \CRANpkg{ape} package.

\code{as.phylo} requires loading the \code{ape} package.
}
\examples{
# A simple example ----------------------------------------------------------

data(fakeexperiment)
data(faketree)
ans <- new_aphylo(fakeexperiment, faketree, "LeafId")
 
# We can visualize it
plot(ans)
}
\seealso{
Other Data management functions: \code{\link{as.phylo}},
  \code{\link{recode_vertices}}
}
