% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.r
\name{new_aphylo_pruner}
\alias{new_aphylo_pruner}
\alias{aphylo_pruner}
\alias{new_aphylo_pruner.matrix}
\alias{new_aphylo_pruner.aphylo}
\alias{new_aphylo_pruner.multiAphylo}
\title{Pointer to \code{pruner}}
\usage{
new_aphylo_pruner(...)

\method{new_aphylo_pruner}{matrix}(edgelist, annotation, types = NULL,
  ...)

\method{new_aphylo_pruner}{aphylo}(x, ...)

\method{new_aphylo_pruner}{multiAphylo}(x, ...)
}
\arguments{
\item{...}{Further arguments passed to the method}

\item{edgelist}{a List two integer vectors.}

\item{annotation}{a list of length \code{N} (annotations).}

\item{types}{An integer vector of types of size \code{N}.}

\item{x}{An object of class \link{aphylo}.}
}
\description{
Creates an external pointer to an object of class \code{aphylo_pruner}. This is mostly
used to compute the model's likelihood function faster by reusing underlying
C++ class objects to store probability matrices and data. This is intended
for internal use only
}
\details{
The underlying implementation of the pruning function is based on the
pruner C++ library that implements Felsenstein's tree pruning algorithm.
See ttps://github.com/USCbiostats/pruner.
}
\examples{
set.seed(1)
x  <- raphylo(10) 
pruner <- new_aphylo_pruner(x)

# Computing loglike
LogLike(
  pruner,
  psi  = c(.10, .20),
  mu_d = c(.90, .80),
  mu_s = c(.10, .05),
  Pi   = .05,
  eta  = c(.90, .80)
  )
}
