% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_aphylo_pruner}
\alias{new_aphylo_pruner}
\title{Pointer to \code{pruner}.}
\usage{
new_aphylo_pruner(edgelist, A, Ntype)
}
\arguments{
\item{edgelist}{a List two integer vectors.}

\item{A}{a list of length \code{N} (annotations).}

\item{Ntype}{An integer vector of types of size \code{N}.}
}
\description{
Pointer to \code{pruner}.
}
\examples{
set.seed(1)
x  <- raphylo(10)
el <- x$tree$edge - 1L
el <- list(el[,1], el[,2])
A  <- lapply(1:19, function(i) with(x, rbind(tip.annotation, node.annotation))[i,]) 
nt <- integer(19)

pruner <- new_aphylo_pruner(el, A, nt)

# Computing loglike
LogLike(pruner, psi = c(.1, .2), mu = c(.1, .05), Pi = .5, eta = c(.9, .8))

}
