% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pies.r
\name{piechart}
\alias{piechart}
\title{A flexible piechart.}
\usage{
piechart(x, labels = names(x), radius = 1, doughnut = 0, origin = c(0,
  0), edges = 200, slice.off = 0, init.angle = 0, last.angle = 360,
  tick.len = 0.1, text.args = list(), segments.args = list(),
  skip.plot.slices = FALSE, add = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector. Values that specify the area of the slices.}

\item{labels}{Character vector of length \code{length(x)}. Passed to
\code{\link[graphics:text]{text}}.}

\item{radius}{Numeric scalar. Radious of the pie.}

\item{doughnut}{Numeric scalar. Radious of the inner circle (doughnut).}

\item{origin}{Numeric vector of length 2. Coordinates of the origin.}

\item{edges}{Numeric scalar. Smoothness of the slices curve.}

\item{slice.off}{Numeric vector. When \code{!=0}, specifies how much to
move the slice away from the origin. When scalar is recycled.}

\item{init.angle}{Numeric scalar. Angle from where to start drawing in degrees.}

\item{last.angle}{Numeric scalar. Angle where to finish drawing in degrees.}

\item{tick.len}{Numeric scalar. Size of the tick marks as \% of the radious.}

\item{text.args}{List. Further arguments passed to \code{\link[graphics:text]{text}}.}

\item{segments.args}{List. Further arguments passed to \code{\link[graphics:segments]{segments}}
when drawing the tickmarks.}

\item{add}{Logical scalar. When \code{TRUE} it is added to the current device.}

\item{...}{Further arguments passed to \code{\link[graphics:polygon]{polygon}}.}
}
\value{
A list with two elements:
\item{slices}{A list of length \code{length(x)} with the coordinates of each
  slice.}
\item{textcoords}{A numeric matrix of size \code{length(x)*2} with 
  coordinates where the labels can be put at.}
\item{alpha0}{A numeric vector of size \code{length(x)} with the starting
degree in radians of the slice.}
\item{alpha1}{A numeric vector of size \code{length(x)} with the ending
degree in radians of the slice.}
}
\description{
A flexible piechart.
}
\examples{
 
vals <- c(1,2,3,10)
piechart(vals, col=grDevices::blues9[5:8], border=NA, doughnut = .5, radius=.75, labs=vals)
piechart(vals, col=grDevices::blues9[3:6], border=NA, doughnut = .3, radius=.5, add=TRUE)
piechart(vals, col=grDevices::blues9[1:4], border=NA, doughnut = .1, radius=.3, add=TRUE)
}
