% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-bis.r
\name{posterior-probabilities}
\alias{posterior-probabilities}
\alias{predict_pre_order}
\alias{predict_pre_order.aphylo_estimates}
\alias{predict_pre_order.aphylo}
\alias{predict_brute_force}
\title{Posterior probabilities based on parameter estimates}
\usage{
predict_pre_order(x, ...)

\method{predict_pre_order}{aphylo_estimates}(x, params = x$par, ...)

\method{predict_pre_order}{aphylo}(x, psi, mu, eta, Pi, ...)

predict_brute_force(atree, psi, mu, Pi)
}
\arguments{
\item{...}{Ignored.}

\item{params}{A numeric vector with the corresponding parameters.}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{mu}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{eta}{Numeric vector of length 2. Annotation bias probabilities (see \code{\link{LogLike}}).}

\item{Pi}{Numeric scalar. Root node probability of having the function (see \code{\link{LogLike}}).}

\item{atree, x}{Either a tree of class \link{aphylo} or an object of class \link{aphylo_estimates}}
}
\description{
The function \code{predict_pre_order} uses a pre-order algorithm to compute the
posterior probabilities, whereas the \code{predict_brute_force} computes posterior
probabilities generating all possible cases.
}
\details{
The function \code{predict_brute_force} is only intended for testing. For predictions
after estimating the model, see \link{predict.aphylo_estimates}.
}
