% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{posterior-probabilities}
\alias{posterior-probabilities}
\alias{predict.aphylo_estimates}
\alias{predict_pre_order}
\alias{predict_pre_order.aphylo_estimates}
\alias{predict_pre_order.aphylo}
\alias{predict_brute_force}
\title{Posterior probabilities based on parameter estimates}
\usage{
\method{predict}{aphylo_estimates}(
  object,
  which.tree = 1:Ntrees(object),
  ids = lapply(Ntip(object)[which.tree], seq_len),
  newdata = NULL,
  loo = TRUE,
  nsamples = 1L,
  centiles = c(0.025, 0.5, 0.975),
  cl = NULL,
  ...
)

predict_pre_order(x, ...)

\method{predict_pre_order}{aphylo_estimates}(
  x,
  params = x$par,
  which.tree = 1:Ntrees(x),
  ids = lapply(Ntip(x)[which.tree], seq_len),
  loo = TRUE,
  nsamples = 1L,
  centiles = c(0.025, 0.5, 0.975),
  ncores = 1L,
  cl = NULL,
  ...
)

\method{predict_pre_order}{aphylo}(x, psi, mu_d, mu_s, eta, Pi, ...)

predict_brute_force(atree, psi, mu_d, mu_s, Pi)
}
\arguments{
\item{which.tree}{Integer scalar. Which tree to include in the prediction.}

\item{ids}{Integer vector. Ids (positions) of the nodes that need to be
predicted.}

\item{newdata}{(optional) An aphylo object.}

\item{loo}{Logical scalar. When \code{loo = TRUE}, predictions are preformed
similar to what a leave-one-out cross-validation scheme would be done
(see \link{predict.aphylo_estimates}).}

\item{nsamples}{Integer scalar. When greater than one, the prediction is done
using a random sample from the MCMC chain. This only works if the model was
fitted using MCMC, of course.}

\item{centiles}{Used together with \code{nsamples}, this indicates the centiles
to be computed from the distribution of outcomes.}

\item{...}{Ignored.}

\item{params}{A numeric vector with the corresponding parameters.}

\item{ncores, cl}{Passed to \code{\link[parallel:makeCluster]{parallel::makeCluster()}}.}

\item{psi}{Numeric vector of length 2. Misclasification probabilities. (see \code{\link{LogLike}}).}

\item{mu_d, mu_s}{Numeric vector of length 2. Gain/loss probabilities (see \code{\link{LogLike}}).}

\item{eta}{Numeric vector of length 2. Annotation bias probabilities (see \code{\link{LogLike}}).}

\item{Pi}{Numeric scalar. Root node probability of having the function (see \code{\link{LogLike}}).}

\item{atree, x, object}{Either a tree of class \link{aphylo} or an object of class \link{aphylo_estimates}}
}
\value{
In the case of the \code{predict} method, a \code{P} column numeric matrix
with values between \eqn{[0,1]} (probabilities).
}
\description{
The function \code{predict_pre_order} uses a pre-order algorithm to compute the
posterior probabilities, whereas the \code{predict_brute_force} computes posterior
probabilities generating all possible cases.
}
\details{
The function \code{predict_brute_force} is only intended for testing. For predictions
after estimating the model, see \link{predict.aphylo_estimates}.

In the case of the parameter \code{loo} (leave-one-out), while making tip-level
predictions, at each leaf the algorithm will drop annotations regarding that
leaf, making its prediction using all the available information except the
one include in such leaf.
}
