% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{sim_annotated_tree}
\alias{sim_annotated_tree}
\title{Simulation of Annotated Phylogenetic Trees}
\usage{
sim_annotated_tree(n = NULL, tree = NULL, P = 1, pars = c(psi0 = 0.05,
  psi1 = 0.05, mu0 = 0.1, mu1 = 0.05, Pi = 0.5))
}
\arguments{
\item{n}{Integer scalar. Number of leafs. If not specified, then}

\item{tree}{An object of class \code{\link[=sim_tree]{phylo_tree}}.}

\item{P}{Integer scalar. Number of functions to generate.}

\item{pars}{Numeric vector of length 5. Parameters of the Data Generating Process
(see \code{\link{LogLike}})}
}
\value{
An object of class \code{\link[=get_offspring]{phylo_offspring}}
}
\description{
Simulation of Annotated Phylogenetic Trees
}
\examples{
# A simple example ----------------------------------------------------------

set.seed(1231)
ans <- sim_annotated_tree(n=500)

}
