% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_fun_on_tree}
\alias{sim_fun_on_tree}
\title{Simulate functions on a ginven tree}
\usage{
sim_fun_on_tree(offspring, noffspring, psi, mu, Pi, P = 1L)
}
\arguments{
\item{offspring}{A List of length \eqn{N} with the set of offspring of
each node.}

\item{noffspring}{An integer vector of length \eqn{N} with the number of
offspring per node.}

\item{psi}{A numeric vector of length 2 (see details).}

\item{mu}{A numeric vector of length 2 (see details).}

\item{Pi}{A numeric vector of length 2 (see details).}

\item{P}{Integer scalar. Number of functions to simulate.}
}
\value{
An matrix of size \code{length(offspring)*P} with values 9, 0 and 1
indicating \code{"no information"}, \code{"no function"} and \code{"function"}.
}
\description{
Simulate functions on a ginven tree
}
\details{
Using the model described in the vignette
\link[=../doc/peeling_phylo.html]{peeling_phylo.html}
}
\examples{
# Example 1 ----------------------------------------------------------------
# Loading the data
data(experiment)
data(tree)
  
# Preprocessing the data
O <- get_offspring(experiment, "LeafId", tree, "NodeId", "ParentId")
    
# Simulating
ans <-
  with(O, sim_fun_on_tree(
      offspring,
      noffspring,
      psi = c(.001, .05) * 0,
      mu = c(.01, .05),
      Pi = c(.5, .5)
  ))
      
# Tabulating results
table(ans)


}
