% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{sim_fun_on_tree}
\alias{sim_fun_on_tree}
\title{Simulate functions on a ginven tree}
\usage{
sim_fun_on_tree(tree, psi, mu, Pi, P = 1L)
}
\arguments{
\item{tree}{An object of class \link[ape:read.tree]{phylo}}

\item{psi}{A numeric vector of length 2 (see details).}

\item{mu}{A numeric vector of length 2 (see details).}

\item{Pi}{A numeric vector of length 2 (see details).}

\item{P}{Integer scalar. Number of functions to simulate.}
}
\value{
An matrix of size \code{length(offspring)*P} with values 9, 0 and 1
indicating \code{"no information"}, \code{"no function"} and \code{"function"}.
}
\description{
Simulate functions on a ginven tree
}
\details{
Using the model described in the vignette
\href{../doc/peeling_phylo.html}{peeling_phylo.html}
}
\examples{
# Example 1 ----------------------------------------------------------------
# We need to simulate a tree
set.seed(1231)
newtree <- sim_tree(1e3)

# Preprocessing the data

# Simulating
ans <- sim_fun_on_tree(
  attr(newtree, "offspring"),
  psi = c(.001, .05),
  mu = c(.01, .05),
  Pi = c(.5, .5)
)

# Tabulating results
table(ans)
}
