% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/data-management.r
\name{sim_tree}
\alias{sim_tree}
\alias{as.phylo.phylo_tree}
\alias{plot.phylo_tree}
\title{Random tree generation}
\usage{
sim_tree(n)

\method{as.phylo}{phylo_tree}(x, ...)

\method{plot}{phylo_tree}(x, y = NULL, tip.color = NULL, ...)
}
\arguments{
\item{n}{Integer scalar. Number of leaf nodes.}

\item{x}{An object of class \code{phylo_tree}.}

\item{...}{Further arguments passed to the method.}

\item{y}{Ignored.}

\item{tip.color}{See \code{\link[ape:plot.phylo]{plot.phylo}}}
}
\value{
An matrix of size \code{n*2 - 2} with column names \code{"offspring"} and
\code{"parent"} representing an edgelist with \code{n*2-1} nodes. A Directed
Acyclic Graph (DAG). Also, includes the following attributes:

\item{offspring}{A list of size \code{n*2 - 1} listing node ith's offspring if any.}
\item{noffspring}{An integer vector of size \code{n*2 - 1} indicating the number of
offspring that each node has.}
}
\description{
By randomly choosing pairs of vertices, this function generates random
trees from bottom to top such that the labels of the nodes follow a 
partial order in their parent-offspring relation, with the parent
always having a lower idlabel than the offspring.
}
\details{
The algorithm was implemented as follows

\enumerate{
  \item Initialize \code{left =[n*2 - 2,...,(n-1)]} and \code{m = n*2 - 2}, and
        initialize the vectors \code{parent} and \code{offspring} to be
        empty.
  \item While \code{length(left) > 1} do:
  \enumerate{
    \item Randomly choose a pair \code{(i, j)} from \code{left}
    \item Add \code{leaf(i)}, \code{leaf(j)} to the tail of \code{offspring},
    \item Decrease \code{m} by 1, and add it two times to the tail of
    \code{parent}.
    \item Remove \code{(i,j)} from \code{leaf} and add \code{m} to its tail.
    \item next
  }
  
}

The \code{\link[ape:rtree]{rtree}} function in the \pkg{ape} package is similar,
although the big difference is in the way the labels are stablished. This later
point is crucial for both \pkg{phylogenetic} and \pkg{ape} as is a key feature
in some (most) of its routines.
}
\examples{
# A very simple example ----------------------------------------------------
set.seed(1223)
newtree <- sim_tree(50)

plot(as.phylo(newtree))

# This is what you would do in igraph --------------------------------------
\dontrun{
g   <- ans
g[] <- as.character(g)
g <- igraph::graph_from_edgelist(g)
plot(g, layout = igraph::layout_with_sugiyama(g)[[2]])
}

# A performance benchmark with ape::rtree ----------------------------------
\dontrun{
microbenchmark::microbenchmark(
ape = rtree(1e3),
  phy = sim_tree(1e3),
unit = "relative"
)
# This is what you would get.
Unit: relative
  expr     min       lq     mean  median       uq      max neval
   ape 14.7598 14.30809 14.30013 16.7217 14.32843 4.754106   100
   phy  1.0000  1.00000  1.00000  1.0000  1.00000 1.000000   100
}
}
